requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
diffusers>=0.12.0
accelerate
Pillow
numpy
opencv-python
gradio>=3.30.0
segmentation-models-pytorch
scikit-image
matplotlib
requests
huggingface_hub
safetensors
peft
controlnet_aux